Decompose multi_domain_ui_routes into 6 focused files
Extract event_routes, route_dispatcher, index_route, show_route,
form_route as includable modules. Phase 3 of HEC-493.
